+++ PinnyNotes.WpfUi/Enums/ToolStates.cs
@@ -1,0 +1,8 @@
+﻿namespace PinnyNotes.WpfUi.Enums;
+
+public enum ToolStates
+{
+    Disabled,
+    Enabled,
+    Favourite
+}