+++ PinnyNotes.WpfUi/Tools/BracketTool.cs
@@ -3,14 +3,14 @@ using System.Linq;
 using System.Windows.Controls;
 
 using PinnyNotes.WpfUi.Commands;
+using PinnyNotes.WpfUi.Enums;
 using PinnyNotes.WpfUi.Properties;
 
 namespace PinnyNotes.WpfUi.Tools;
 
 public class BracketTool : BaseTool, ITool
 {
-    public bool IsEnabled => ToolSettings.Default.BracketToolEnabled;
-    public bool IsFavourite => ToolSettings.Default.BracketToolFavourite;
+    public ToolStates State => (ToolStates)Settings.Default.BracketToolState;
 
     private static char[] _openingBrackets = {'(', '{', '['};
     private static char[] _closingBrackets = { ')', '}', ']' };