+++ PinnyNotes.WpfUi/Tools/GibberishTool.cs
@@ -2,14 +2,14 @@
 using System.Windows.Controls;
 
 using PinnyNotes.WpfUi.Commands;
+using PinnyNotes.WpfUi.Enums;
 using PinnyNotes.WpfUi.Properties;
 
 namespace PinnyNotes.WpfUi.Tools;
 
 public class GibberishTool : BaseTool, ITool
 {
-    public bool IsEnabled => ToolSettings.Default.GibberishToolEnabled;
-    public bool IsFavourite => ToolSettings.Default.GibberishToolFavourite;
+    public ToolStates State => (ToolStates)Settings.Default.GibberishToolState;
 
     public enum ToolActions
     {