+++ PinnyNotes.WpfUi/Tools/RemoveTool.cs
@@ -4,14 +4,14 @@ using System.Text;
 using System.Windows.Controls;
 
 using PinnyNotes.WpfUi.Commands;
+using PinnyNotes.WpfUi.Enums;
 using PinnyNotes.WpfUi.Properties;
 
 namespace PinnyNotes.WpfUi.Tools;
 
 public class RemoveTool : BaseTool, ITool
 {
-    public bool IsEnabled => ToolSettings.Default.RemoveToolEnabled;
-    public bool IsFavourite => ToolSettings.Default.RemoveToolFavourite;
+    public ToolStates State => (ToolStates)Settings.Default.RemoveToolState;
 
     private string? _selectedText = null;