+++ PinnyNotes.WpfUi/Tools/SplitTool.cs
@@ -2,6 +2,7 @@
 using System.Windows.Controls;
 
 using PinnyNotes.WpfUi.Commands;
+using PinnyNotes.WpfUi.Enums;
 using PinnyNotes.WpfUi.Properties;
 
 namespace PinnyNotes.WpfUi.Tools;
@@ -10,8 +11,7 @@ namespace PinnyNotes.WpfUi.Tools;
 
 public class SplitTool : BaseTool, ITool
 {
-    public bool IsEnabled => ToolSettings.Default.SplitToolEnabled;
-    public bool IsFavourite => ToolSettings.Default.SplitToolFavourite;
+    public ToolStates State => (ToolStates)Settings.Default.SplitToolState;
 
     private string? _selectedText = null;