+++ PinnyNotes.WpfUi/Tools/TrimTool.cs
@@ -2,14 +2,14 @@
 using System.Windows.Controls;
 
 using PinnyNotes.WpfUi.Commands;
+using PinnyNotes.WpfUi.Enums;
 using PinnyNotes.WpfUi.Properties;
 
 namespace PinnyNotes.WpfUi.Tools;
 
 public class TrimTool : BaseTool, ITool
 {
-    public bool IsEnabled => ToolSettings.Default.TrimToolEnabled;
-    public bool IsFavourite => ToolSettings.Default.TrimToolFavourite;
+    public ToolStates State => (ToolStates)Settings.Default.TrimToolState;
 
     public enum ToolActions
     {